bump version and release beta with misc fixes

+++ cumulus_process/process.py
@@ -9,6 +9,7 @@ import cumulus_process.s3 as s3
 from cumulus_process.loggers import getLogger
 from cumulus_process.cli import cli
 from cumulus_process.handlers import activity
+from run_cumulus_task import run_cumulus_task
 
 logger = getLogger(__name__)
 
@@ -56,6 +57,7 @@ class Process(object):
 
     def __init__(self, input, path='./', config={}, **kwargs):
         """ Initialize a Process with input filenames and optional kwargs """
+        # local work directory files will be stored
         self.path = path
         self.config = config
         self.kwargs = kwargs
@@ -85,6 +87,7 @@ class Process(object):
                     outfiles.append(f)
                 else:
                     outfiles.append(s3.download(f, path=self.path))
+        return outfiles
 
     def fetch_all(self, remote=False):
         """ Download all files in remote_in """
@@ -104,13 +107,13 @@ class Process(object):
 
     def upload_output_files(self):
         """ Uploads all self.outputs """
-        return [self.upload_file(f) for f in self.outputs]
+        return [self.upload_file(f) for f in self.output]
 
     def clean_input(self):
         """ Remove input files """
         self.logger.info('Cleaning local input files')
         for f in self.input:
-            if os.path.exists(f):
+            if os.path.exists(f) and f not in self.output:
                 os.remove(f)
 
     def clean_output(self):
@@ -207,19 +210,38 @@ class Process(object):
         return os.path.splitext(os.path.basename(filename))[0]
 
     # ## Handlers
-
     @classmethod
     def handler(cls, event, context=None):
-        process = cls(**event)
-        return process.process()
+        """ General event handler """
+        return cls.run(**event)
+
+    @classmethod
+    def cumulus_handler(cls, event, context=None):
+        """ General event handler using Cumulus messaging (cumulus-message-adapter) """
+        outputs = run_cumulus_task(cls.handler, event, context)
+        files = []
+        for f in outputs:
+            uri = s3.uri_parser(f)
+            files.append({
+                'filename': f,
+                'name': uri['filename'],
+                'bucket': uri['bucket']
+            })
+        return files
 
     @classmethod
     def cli(cls):
         cli(cls)
 
     @classmethod
-    def activity(cls, arn):
-        activity(cls.handler, arn=arn)
+    def activity(cls, arn=os.getenv('ACTIVITY_ARN')):
+        """ Run an AWS activity for a step function """
+        activity(cls.handler, arn)
+
+    @classmethod
+    def cumulus_activity(cls, arn=os.getenv('ACTIVITY_ARN')):
+        """ Run an activity using Cumulus messaging (cumulus-message-adapter) """
+        activity(cls.cumulus_handler, arn)
 
     @classmethod
     def run(cls, *args, **kwargs):
